# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: School ID App API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List students
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: grade
          in: query
          schema:
            type: string
        - name: className
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of students with current state
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentWithState"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentBody"
      responses:
        "201":
          description: Student created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get a student with full profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student with state and today's timeline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
        "404":
          description: Not found
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentBody"
      responses:
        "200":
          description: Student updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /students/lookup/{qrCode}:
    get:
      operationId: lookupStudentByQr
      tags: [students]
      summary: Lookup a student by QR code or student ID
      parameters:
        - name: qrCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Student with current state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentWithState"
        "404":
          description: Not found
...
components:
  schemas:
...
    Student:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        grade:
          type: string
        className:
          type: string
        photoUrl:
          type: ["string", "null"]
        qrCode:
          type: string
        isActive:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - studentId
        - firstName
        - lastName
        - grade
        - className
        - qrCode
        - isActive
        - createdAt
        - updatedAt
    StudentWithState:
      allOf:
        - $ref: "#/components/schemas/Student"
        - type: object
          properties:
            currentState:
              type: string
            lastSeenAt:
              type: ["string", "null"]
            lastSeenLocation:
              type: ["string", "null"]
          required:
            - currentState
    StudentProfile:
      allOf:
        - $ref: "#/components/schemas/StudentWithState"
        - type: object
          properties:
            todayTimeline:
              type: array
              items:
                $ref: "#/components/schemas/ScanEvent"
            behaviorSummary:
              $ref: "#/components/schemas/BehaviorSummary"
          required:
            - todayTimeline
            - behaviorSummary
    BehaviorSummary:
      type: object
      properties:
        totalMerits:
          type: integer
        totalDemerits:
          type: integer
        recentLogs:
          type: array
          items:
            $ref: "#/components/schemas/BehaviorLog"
      required:
        - totalMerits
        - totalDemerits
        - recentLogs
    CreateStudentBody:
      type: object
      properties:
        studentId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        grade:
          type: string
        className:
          type: string
        photoUrl:
          type: ["string", "null"]
      required:
        - studentId
        - firstName
        - lastName
        - grade
        - className
    UpdateStudentBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        grade:
          type: string
        className:
          type: string
        photoUrl:
          type: ["string", "null"]
        isActive:
          type: integer
    ScanType:
      type: string
      enum:
        - gate_in
        - gate_out
        - checkout
        - class
        - event
        - assembly
        - activity
        - detention
        - club
...
    ScanEvent:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        scannedById:
          type: ["integer", "null"]
        scanType:
          $ref: "#/components/schemas/ScanType"
        location:
          type: ["string", "null"]
        activityId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        studentName:
          type: ["string", "null"]
      required:
        - id
        - studentId
        - scanType
        - createdAt
...
    BehaviorLog:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        categoryId:
          type: ["integer", "null"]
        type:
          type: string
        points:
          type: integer
        note:
          type: ["string", "null"]
        loggedById:
          type: ["integer", "null"]
        createdAt:
          type: string
        categoryName:
          type: ["string", "null"]
      required:
        - id
        - studentId
        - type
        - points
        - createdAt
...